Move command deduplication feature descriptors into the experimental features message.
This indicates the fact that those feature descriptors are for testing purposes only.

Breaks-protobuf: true
